+++ src/mahony.rs
@@ -1,7 +1,7 @@
 #![allow(non_snake_case)]
 #![allow(clippy::many_single_char_names)]
 
-use crate::ahrs::Ahrs;
+use crate::ahrs::{Ahrs, AhrsError};
 use core::hash;
 use nalgebra::{Quaternion, Scalar, UnitQuaternion, Vector2, Vector3};
 use simba::simd::{SimdRealField as RealField, SimdRealField, SimdValue};
@@ -205,7 +205,7 @@ impl<N: simba::scalar::RealField> Ahrs<N> for Mahony<N> {
         gyroscope: &Vector3<N>,
         accelerometer: &Vector3<N>,
         magnetometer: &Vector3<N>,
-    ) -> Result<&UnitQuaternion<N>, &str> {
+    ) -> Result<&UnitQuaternion<N>, AhrsError> {
         let q = self.quat.as_ref();
 
         let zero: N = nalgebra::zero();
@@ -216,7 +216,7 @@ impl<N: simba::scalar::RealField> Ahrs<N> for Mahony<N> {
         let accel = match accelerometer.try_normalize(zero) {
             Some(n) => n,
             None => {
-                return Err("Accelerometer norm divided by zero.");
+                return Err(AhrsError::DivByZero);
             }
         };
 
@@ -224,7 +224,7 @@ impl<N: simba::scalar::RealField> Ahrs<N> for Mahony<N> {
         let mag = match magnetometer.try_normalize(zero) {
             Some(n) => n,
             None => {
-                return Err("Magnetometer norm divided by zero.");
+                return Err(AhrsError::DivByZero);
             }
         };
 
@@ -274,7 +274,7 @@ impl<N: simba::scalar::RealField> Ahrs<N> for Mahony<N> {
         &mut self,
         gyroscope: &Vector3<N>,
         accelerometer: &Vector3<N>,
-    ) -> Result<&UnitQuaternion<N>, &str> {
+    ) -> Result<&UnitQuaternion<N>, AhrsError> {
         let q = self.quat.as_ref();
 
         let zero: N = nalgebra::zero();
@@ -285,7 +285,7 @@ impl<N: simba::scalar::RealField> Ahrs<N> for Mahony<N> {
         let accel = match accelerometer.try_normalize(zero) {
             Some(n) => n,
             None => {
-                return Err("Accelerometer norm divided by zero.");
+                return Err(AhrsError::DivByZero);
             }
         };